[GlobalDCE] Add !llvm.used.metadata support for conditionally used global variables
As part of the optimization work to emit dead-strippable symbols for Swift and
enable GlobalDCE to remove those symbols when possible, we need to mark some
Swift symbols as conditionally used based on a set of other symbols. Background
info can be found here: https://bugs.swift.org/browse/SR-14509.

This diff adds a module-level name metadata !llvm.used.conditional, which
contains a list of metadata triplets, that describes the conditions under which
we allow removal of globals even if they are mentioned in @llvm.used. See the
included LangRef changes to details. The implementation is a strict addition,
it does not change any semantics and does not affect IR or passes in any way,
unless !llvm.used.conditional is present in the input IR.